[All] fixes Ticket #5232 - broken sort-by-size when files are bigger than 4GB. tested to 11.5GB files
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@16014 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/StringUtils.h
@@ -52,7 +52,7 @@ public:
   static void JoinString(const CStdStringArray &strings, const CStdString& delimiter, CStdString& result);
   static int SplitString(const CStdString& input, const CStdString& delimiter, CStdStringArray &results, unsigned int iMaxStrings = 0);
   static int FindNumber(const CStdString& strInput, const CStdString &strFind);
-  static int AlphaNumericCompare(const char *left, const char *right);
+  static long long AlphaNumericCompare(const char *left, const char *right);
   static long TimeStringToSeconds(const CStdString &timeString);
   static void RemoveCRLF(CStdString& strLine);
   static void SecondsToTimeString( long lSeconds, CStdString& strHMS, TIME_FORMAT format = TIME_FORMAT_GUESS);
